local/lambda/requirements.txt
requests
boto3
google-generativeai